requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-core==0.1.10
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2
openai==1.3.7
numpy==1.24.3
